VCPDesktop/requirements.txt
websockets
Pillow
mss
requests
